BACKPORT: releasetools: Generate recovery-from-boot.p with bsdiff if necessary
Original commit message:
"If TARGET_NOT_USE_GZIP_RECOVERY_RAMDISK is true, we want to use bsdiff

to create recovery-from-boot.p, otherwise there are high chances that

imgdiff will fail. Currently this is done only when running make, but

not when we re-create the patch from releasetools, so do it.

Also, since recovery-resource.dat is used as bonus data for imgdiff,

don't build it if we are going to use bsdiff.

Change-Id: I93a662a358ee79f56b8acd4329eedd166a176c66"